+++ _build/resolvers/sample_settings.php
@@ -22,7 +22,6 @@ if (!function_exists('getResourceMap')) {
 
 switch ($options[xPDOTransport::PACKAGE_ACTION]) {
     case xPDOTransport::ACTION_INSTALL:
-    case xPDOTransport::ACTION_UPGRADE:
         if (isset($options['install_resources']) && empty($options['install_resources'])) return true;
         
         $resourceMap =  getResourceMap();
@@ -59,6 +58,8 @@ switch ($options[xPDOTransport::PACKAGE_ACTION]) {
             }
         }
 
+        break;
+    case xPDOTransport::ACTION_UPGRADE:
         break;
     case xPDOTransport::ACTION_UNINSTALL:
         break;