+++ _build/setup.options.php
@@ -1,6 +1,8 @@
 <?php
 
-$output = '<label for="install_resources">Install sample content:</label>
+$output = 'After installation, please turn ON system setting <strong>friendly_urls</strong>.
+<br />
+<label for="install_resources">Install sample content:</label>
 <select name="install_resources" id="install_resources">
     <option value="">No</option>
     <option value="yes">Yes</option>