+++ core/components/flatso/docs/changelog.md
@@ -1,2 +1,2 @@
 # Flatso 1.0.0
-- Basic Blog with 2.4 Dependencies and System Config settings`
+- Basic Blog with 2.4 Dependencies and System Config settings